feat: app passed stress test (29 KOs only)
The issue from previous commit stemmed from resource limitation, not allowing to initialize worker threads.

Another idea was to make use of forks (multiprocessing in python), but it is not compatible with gevent, apparently.

The decision made was to split the project in two: this one (async, without threads) and another one using only forks.

Since the last day to submit projects to the contest is tomorrow, i won't submit the second project, but will work on it separately.

It is important to say that the code from the previous commit did work, so it can be used for further reference, but not in cases where resources are limited such as this one (with less than a cpu core for each api).